refactor(host-application): split task workflow lifecycle support modules
