+++ apps/desktop/src-tauri/crates/host-application/src/app_service/task_workflow/implementation_reset_service.rs
@@ -1,8 +1,9 @@
 use super::{
-    lifecycle_support::{
+    cleanup_plans::{
         derive_reset_implementation_status, ensure_task_reset_status_allowed,
-        with_reset_cleanup_progress, BranchCleanupPlan, TaskActivityGuard, WorktreeCleanupPlan,
+        with_reset_cleanup_progress, BranchCleanupPlan, WorktreeCleanupPlan,
     },
+    task_activity_guard::TaskActivityGuard,
     task_context::LoadedTaskContext,
 };
 use crate::app_service::service_core::AppService;
@@ -27,138 +28,17 @@ impl<'a> ImplementationResetService<'a> {
         let task_id = context.task.id.clone();
         ensure_task_reset_status_allowed(&context.task)?;
 
-        let sessions = self
-            .service
-            .agent_sessions_list(context.repo.repo_path.as_str(), task_id.as_str())?;
-        TaskActivityGuard::new(self.service).ensure_no_active_task_reset_runs(
-            context.repo.repo_path.as_str(),
-            task_id.as_str(),
-            &sessions,
-        )?;
+        let cleanup_plan = self.build_cleanup_plan(&context)?;
+        let mut cleanup_progress = ResetCleanupProgress::default();
 
-        let rollback_status = derive_reset_implementation_status(&context.task);
-        let branch_prefix = self
-            .service
-            .config_store
-            .repo_config(&context.repo.repo_path)?
-            .branch_prefix;
-        let branch_plan = BranchCleanupPlan::for_task(
-            self.service,
+        self.execute_cleanup(&context, &cleanup_plan, &mut cleanup_progress)?;
+        self.clear_reset_metadata(context.repo_dir(), task_id.as_str(), &cleanup_progress)?;
+        let updated = self.apply_reset_status(
             context.repo_dir(),
-            branch_prefix.as_str(),
             task_id.as_str(),
+            cleanup_plan.rollback_status,
+            &cleanup_progress,
         )?;
-        let worktree_plan = WorktreeCleanupPlan::for_task_sessions(
-            self.service,
-            context.repo.repo_path.as_str(),
-            task_id.as_str(),
-            branch_prefix.as_str(),
-            &sessions,
-            true,
-        )?;
-
-        self.service
-            .stop_dev_servers_for_task(context.repo.repo_path.as_str(), task_id.as_str())?;
-        let mut removed_worktrees = Vec::new();
-        let mut deleted_branches = Vec::new();
-
-        for worktree_path in worktree_plan.paths() {
-            if let Err(error) = self
-                .service
-                .git_remove_worktree(context.repo.repo_path.as_str(), worktree_path, true)
-                .with_context(|| {
-                    format!("Failed to remove implementation worktree {worktree_path}")
-                })
-            {
-                return Err(with_reset_cleanup_progress(
-                    error,
-                    &removed_worktrees,
-                    &deleted_branches,
-                ));
-            }
-            removed_worktrees.push(worktree_path.clone());
-        }
-
-        if let Err(error) = branch_plan.ensure_unused_by_worktrees(self.service, context.repo_dir())
-        {
-            return Err(with_reset_cleanup_progress(
-                error,
-                &removed_worktrees,
-                &deleted_branches,
-            ));
-        }
-
-        for branch_name in branch_plan.names() {
-            if let Err(error) = self
-                .service
-                .git_delete_local_branch(
-                    context.repo.repo_path.as_str(),
-                    branch_name.as_str(),
-                    true,
-                )
-                .with_context(|| format!("Failed to delete implementation branch {branch_name}"))
-            {
-                return Err(with_reset_cleanup_progress(
-                    error,
-                    &removed_worktrees,
-                    &deleted_branches,
-                ));
-            }
-            deleted_branches.push(branch_name.clone());
-        }
-
-        self.service
-            .task_store
-            .clear_agent_sessions_by_roles(context.repo_dir(), task_id.as_str(), &["build", "qa"])
-            .with_context(|| format!("Failed to clear builder and QA sessions for {task_id}"))
-            .map_err(|error| {
-                with_reset_cleanup_progress(error, &removed_worktrees, &deleted_branches)
-            })?;
-        self.service
-            .task_store
-            .clear_qa_reports(context.repo_dir(), task_id.as_str())
-            .with_context(|| format!("Failed to clear QA reports for {task_id}"))
-            .map_err(|error| {
-                with_reset_cleanup_progress(error, &removed_worktrees, &deleted_branches)
-            })?;
-        self.service
-            .task_store
-            .set_pull_request(context.repo_dir(), task_id.as_str(), None)
-            .with_context(|| format!("Failed to clear linked pull request for {task_id}"))
-            .map_err(|error| {
-                with_reset_cleanup_progress(error, &removed_worktrees, &deleted_branches)
-            })?;
-        self.service
-            .task_store
-            .set_direct_merge_record(context.repo_dir(), task_id.as_str(), None)
-            .with_context(|| format!("Failed to clear direct merge metadata for {task_id}"))
-            .map_err(|error| {
-                with_reset_cleanup_progress(error, &removed_worktrees, &deleted_branches)
-            })?;
-
-        let updated = self
-            .service
-            .task_store
-            .update_task(
-                context.repo_dir(),
-                task_id.as_str(),
-                UpdateTaskPatch {
-                    title: None,
-                    description: None,
-                    notes: None,
-                    status: Some(rollback_status),
-                    priority: None,
-                    issue_type: None,
-                    ai_review_enabled: None,
-                    labels: None,
-                    assignee: None,
-                    parent_id: None,
-                },
-            )
-            .with_context(|| format!("Failed to reset implementation for {task_id}"))
-            .map_err(|error| {
-                with_reset_cleanup_progress(error, &removed_worktrees, &deleted_branches)
-            })?;
 
         if let Some(index) = context
             .repo
@@ -174,4 +54,205 @@ impl<'a> ImplementationResetService<'a> {
 
         Ok(self.service.enrich_task(updated, &context.repo.tasks))
     }
+
+    fn build_cleanup_plan(&self, context: &LoadedTaskContext) -> Result<ImplementationResetPlan> {
+        let task_id = context.task.id.as_str();
+        let sessions = self
+            .service
+            .agent_sessions_list(context.repo.repo_path.as_str(), task_id)?;
+        TaskActivityGuard::new(self.service).ensure_no_active_task_reset_runs(
+            context.repo.repo_path.as_str(),
+            task_id,
+            &sessions,
+        )?;
+        let rollback_status = derive_reset_implementation_status(&context.task);
+        let branch_prefix = self
+            .service
+            .config_store
+            .repo_config(&context.repo.repo_path)?
+            .branch_prefix;
+        let branch_plan = BranchCleanupPlan::for_task(
+            self.service,
+            context.repo_dir(),
+            branch_prefix.as_str(),
+            task_id,
+        )?;
+        let worktree_plan = WorktreeCleanupPlan::for_task_sessions(
+            self.service,
+            context.repo.repo_path.as_str(),
+            task_id,
+            branch_prefix.as_str(),
+            &sessions,
+            true,
+        )?;
+
+        Ok(ImplementationResetPlan {
+            rollback_status,
+            branch_plan,
+            worktree_plan,
+        })
+    }
+
+    fn execute_cleanup(
+        &self,
+        context: &LoadedTaskContext,
+        cleanup_plan: &ImplementationResetPlan,
+        cleanup_progress: &mut ResetCleanupProgress,
+    ) -> Result<()> {
+        let task_id = context.task.id.as_str();
+        self.service
+            .stop_dev_servers_for_task(context.repo.repo_path.as_str(), task_id)?;
+
+        for worktree_path in cleanup_plan.worktree_plan.paths() {
+            if let Err(error) = self
+                .service
+                .git_remove_worktree(context.repo.repo_path.as_str(), worktree_path, true)
+                .with_context(|| {
+                    format!("Failed to remove implementation worktree {worktree_path}")
+                })
+            {
+                return Err(with_reset_cleanup_progress(
+                    error,
+                    &cleanup_progress.removed_worktrees,
+                    &cleanup_progress.deleted_branches,
+                ));
+            }
+            cleanup_progress
+                .removed_worktrees
+                .push(worktree_path.clone());
+        }
+
+        if let Err(error) = cleanup_plan
+            .branch_plan
+            .ensure_unused_by_worktrees(self.service, context.repo_dir())
+        {
+            return Err(with_reset_cleanup_progress(
+                error,
+                &cleanup_progress.removed_worktrees,
+                &cleanup_progress.deleted_branches,
+            ));
+        }
+
+        for branch_name in cleanup_plan.branch_plan.names() {
+            if let Err(error) = self
+                .service
+                .git_delete_local_branch(
+                    context.repo.repo_path.as_str(),
+                    branch_name.as_str(),
+                    true,
+                )
+                .with_context(|| format!("Failed to delete implementation branch {branch_name}"))
+            {
+                return Err(with_reset_cleanup_progress(
+                    error,
+                    &cleanup_progress.removed_worktrees,
+                    &cleanup_progress.deleted_branches,
+                ));
+            }
+            cleanup_progress.deleted_branches.push(branch_name.clone());
+        }
+
+        Ok(())
+    }
+
+    fn clear_reset_metadata(
+        &self,
+        repo_dir: &std::path::Path,
+        task_id: &str,
+        cleanup_progress: &ResetCleanupProgress,
+    ) -> Result<()> {
+        self.service
+            .task_store
+            .clear_agent_sessions_by_roles(repo_dir, task_id, &["build", "qa"])
+            .with_context(|| format!("Failed to clear builder and QA sessions for {task_id}"))
+            .map_err(|error| {
+                with_reset_cleanup_progress(
+                    error,
+                    &cleanup_progress.removed_worktrees,
+                    &cleanup_progress.deleted_branches,
+                )
+            })?;
+        self.service
+            .task_store
+            .clear_qa_reports(repo_dir, task_id)
+            .with_context(|| format!("Failed to clear QA reports for {task_id}"))
+            .map_err(|error| {
+                with_reset_cleanup_progress(
+                    error,
+                    &cleanup_progress.removed_worktrees,
+                    &cleanup_progress.deleted_branches,
+                )
+            })?;
+        self.service
+            .task_store
+            .set_pull_request(repo_dir, task_id, None)
+            .with_context(|| format!("Failed to clear linked pull request for {task_id}"))
+            .map_err(|error| {
+                with_reset_cleanup_progress(
+                    error,
+                    &cleanup_progress.removed_worktrees,
+                    &cleanup_progress.deleted_branches,
+                )
+            })?;
+        self.service
+            .task_store
+            .set_direct_merge_record(repo_dir, task_id, None)
+            .with_context(|| format!("Failed to clear direct merge metadata for {task_id}"))
+            .map_err(|error| {
+                with_reset_cleanup_progress(
+                    error,
+                    &cleanup_progress.removed_worktrees,
+                    &cleanup_progress.deleted_branches,
+                )
+            })?;
+
+        Ok(())
+    }
+
+    fn apply_reset_status(
+        &self,
+        repo_dir: &std::path::Path,
+        task_id: &str,
+        rollback_status: host_domain::TaskStatus,
+        cleanup_progress: &ResetCleanupProgress,
+    ) -> Result<TaskCard> {
+        self.service
+            .task_store
+            .update_task(
+                repo_dir,
+                task_id,
+                UpdateTaskPatch {
+                    title: None,
+                    description: None,
+                    notes: None,
+                    status: Some(rollback_status),
+                    priority: None,
+                    issue_type: None,
+                    ai_review_enabled: None,
+                    labels: None,
+                    assignee: None,
+                    parent_id: None,
+                },
+            )
+            .with_context(|| format!("Failed to reset implementation for {task_id}"))
+            .map_err(|error| {
+                with_reset_cleanup_progress(
+                    error,
+                    &cleanup_progress.removed_worktrees,
+                    &cleanup_progress.deleted_branches,
+                )
+            })
+    }
+}
+
+struct ImplementationResetPlan {
+    rollback_status: host_domain::TaskStatus,
+    branch_plan: BranchCleanupPlan,
+    worktree_plan: WorktreeCleanupPlan,
+}
+
+#[derive(Default)]
+struct ResetCleanupProgress {
+    removed_worktrees: Vec<String>,
+    deleted_branches: Vec<String>,
 }
