+++ apps/desktop/src-tauri/crates/host-application/src/app_service/task_workflow/mod.rs
@@ -1,11 +1,12 @@
 mod approval_service;
+mod cleanup_plans;
 mod document_service;
 mod implementation_reset_service;
-mod lifecycle_support;
 mod qa_service;
 mod session_service;
+mod task_activity_guard;
 mod task_context;
 mod task_deletion_service;
 mod task_service;
 
-pub(crate) use lifecycle_support::{normalize_path_for_comparison, normalize_path_key};
+pub(crate) use cleanup_plans::{normalize_path_for_comparison, normalize_path_key};
